requirements.txt
django==3.2.3
django-filter==2.4.0
django-allauth==0.44.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
django-dynamic-preferences==1.10.1
psycopg2-binary==2.8.6
django-tenants==3.3.1
django-storages==1.11.1
django-phonenumber-field==5.1.0
phonenumbers==8.12.24
jsonfield==3.1.0
cerberus==1.3.4
pillow==8.2.0
simplejson==3.17.2
boto3==1.17.84
awscli==1.19.84
